Updates to Player Movement, Art, and Models
Added the bool isInAMenu to the PlayerMovement script and referenced isInAMenu in the PuzzleInteraction script, toggling it when the UI is toggled.

This is an updated version of all that is in the Unity project, including the art and model things that Ryan added.

Did not include the things that it said would get deleted (just in case)