Reedition of PuzzleReadyCanvas
So, apparently, some time between our class meeting and now, the Puzzle Ready Canvas got deleted. Rest assured, I was able to reload an older version of the project was able to get it back. This is the final change to the main branch before we start dev up again.