Started Menu Building (Main and Pause)
Started the PauseMenu and MainMenu scripts, making a MainMixer for volume control, and a MainMenuCanvas. The MainMenuCanvas has the MainMenu script and has the MainMenuElements and OptionsMenuElements hooked up to it. The MainMenuCanvas also has a VolumeSlider under the OptionsMenuElements.
The pause menu has a transparent gray button that uses the Resume() function (when they player clicks outside the base area, it closes the pause menu). Then, removed the pause menu from the MainMenu scene.
Made the MainMenuCanvas into a prefab. Made the PauseMenuCanvas into a prefab. Made a UI folder in the Prefabs folder. Made a UI folder in the scripts folder.