Come Clean Up and Polish
Removed a reference to PuzzleUI from PlayerMovement because it was never used. Removed the setting of speed from PuzzleUI because speed doesn't need to be changed at all after the isInAMenu bool was implemented. Commented out, but didn't delete the Rewired example from PlayerMovement.
Tried to fix audio issues. I think the problem is that the AudioClip PuzzleStart does not have an AudioSource, so it's not going through the MainMixer. Not sure yet how to fix it.