Made Door 1 Switch Case
Moved all the code from GameManager to the LevelManager. Implemented the door 1 switch case, checking if the tutorial puzzles are complete to open the door. Assigned it to a temp cube in the scene.
Prevents puzzle UI from showing up on the main menu.
Make a SleepingBayBehavior script. Added BoxCollider to the sleeping bay which becomes enabled when all five puzzles are complete. When the player then interacts with the sleeping bay, the powering down animation plays, waits 10 seconds, and then ups the cycle and plays the powering up animation (animation are yet to exist).
Removed other scenes from the build settings.