Implemented the FirstDoor, No Name Changes
Made the isOpening bool, set up the conditions in the animator controller 'Door Animator', put the door model in the scene, added the DoorBehavior script to the door, put the door in the door list in the LevelMananger.
Also set it so that the sleeping bay doors don't loop after opening.