Save System Updates
I deleted CanvasPuzzles 'cause it got messed up and it was discovered that it wasn't be used any way.
Tried to fix the sleeping bay animation. I got them opening again, but it won't stop looping.
Re-hooked up the menu and pause save/load buttons because somehow they got undone -_- Very annoying
Set up loops to save the conditions of the puzzles and the doors (isComplete, isOpen). Using them in loading the game is not implemented yet.
I set it up so that the player can load into the sleeping bay in the lists in the LevelManager. For more information on how it works, go to the Game Manager Jamboard, slide 5. It needs some adjustments to make it so the position of the sleeping bay is the center of the bay, rather than a corner. Other than that I think it runs as expected, but it needs much more testing to be sure.