Added Camera to Prevent Error
Added a Virtual Camera called Door1CutScene to the LevelManager's CutSceneCams list, inside CameraHolder in the hierarchy.  The cut scene doesn't actually show the door opening currently, but it doesn't throw any errors, which was the current goal. Commented out the sound effect when the door opens, because we don't have that yet, I think.