Cynthia's Updates 1
Added the CreditsManager script and its code. Hooked it up and it is functional.
Locked the cursor when not in a menu. Doesn't work with the ESC key though. I think it's Rewired because the only difference between the buttons and ESC is Rewired. They use the exact same functions.
Moved the box collider for the second puzzle.