Added End of Game Condition and Credits Update
The game will now go to the credits scene if the player hits 'E' while in the Captain's Chair trigger box collider.
The skip text in the credits scene is no longer blurry. The credits scene also have the background music in it (attacked to the background raw image on the canvas.