5-19-21 Later Update
Took screenshots of the jamboard solutions and put them in a folder in the project. Finished the 2nd floor door animations and they all work as intended. (Used Destroy() on the Fake Wall). Replaced some of the remaining Door Buttons with Elevator Console Placeholders, because they look nicer. Edited the task names again. Added code (if we want to keep it) so that Bitbot's Speed is increased and turntime decreased after the 1st cycle, as a bonus. It's currently commented out, but it's in the LevelManager script, beginning of the 2nd cycle.